# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/club": {
      "get": {
        "operationId": "ClubController_searchClubs",
        "parameters": [
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Container para os filtros de busca de clubes.",
            "schema": {
              "$ref": "#/components/schemas/SearchClubsFilterDto"
            }
          },
          {
            "name": "pagination",
            "required": false,
            "in": "query",
            "description": "Container para os parâmetros de paginação.",
            "schema": {
              "$ref": "#/components/schemas/PaginationDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clubes retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedClubDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Busca e lista clubes com paginação e filtros",
        "tags": [
          "Clubes"
        ]
      },
      "post": {
        "operationId": "ClubController_createClub",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Clube criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClubOutputDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cria um novo clube",
        "tags": [
          "Clubes"
        ]
      }
    },
    "/club/{id}": {
      "get": {
        "operationId": "ClubController_getClubInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do clube retornados com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              }
            }
          },
          "404": {
            "description": "Clube não encontrado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtém informações detalhadas de um clube específico",
        "tags": [
          "Clubes"
        ]
      }
    },
...
    "/club-management/my-club": {
      "get": {
        "operationId": "ClubManagementController_getMyClubInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dados do clube retornados com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtém as informações do clube do diretor autenticado",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    },
    "/club-management/{clubId}": {
      "patch": {
        "operationId": "ClubManagementController_updateClub",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Clube atualizado com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualiza as informações de um clube específico que o usuário dirige",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    },
    "/club-management/my-club/enrollments": {
      "get": {
        "operationId": "ClubManagementController_listEnrollments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de solicitações pendentes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentRequestDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todas as solicitações de matrícula para o clube do usuário logado.",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    },
    "/club-management/{clubId}/enrollments/pending": {
      "get": {
        "operationId": "ClubManagementController_listPending",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitações pendentes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListPendingEnrollmentsOutputDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista as solicitações de matrícula pendentes para um clube específico",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    },
    "/club-management/enrollments/{enrollmentId}/approve": {
      "post": {
        "operationId": "ClubManagementController_approve",
        "parameters": [
          {
            "name": "enrollmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Matrícula aprovada com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Aprova uma solicitação de matrícula",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    },
    "/club-management/enrollments/{enrollmentId}/reject": {
      "post": {
        "operationId": "ClubManagementController_reject",
        "parameters": [
          {
            "name": "enrollmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Matrícula rejeitada com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Rejeita uma solicitação de matrícula",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    },
    "/club-management/my-club/members": {
      "get": {
        "operationId": "ClubManagementController_listMembersOfMyClub",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de membros ativos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubMemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todos os membros ativos do meu clube",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    },
    "/club-management/membership/{membershipId}/revoke": {
      "post": {
        "operationId": "ClubManagementController_removeMember",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Afiliação do membro revogada com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Revoga a afiliação de um membro do clube",
        "tags": [
          "Gestão de Clube (Diretor)"
        ]
      }
    }
  },
  "info": {
    "title": "API NCFCA",
    "description": "Documentação da API para o sistema NCFCA. Esta é a única fonte de verdade para os contratos da API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Entre com o token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "SearchClubsFilterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Filtra clubes pelo nome (busca parcial, case-insensitive).",
            "example": "Oradores"
          },
          "city": {
            "type": "string",
            "description": "Filtra clubes pela cidade (busca exata, case-insensitive).",
            "example": "Brasília"
          },
          "state": {
            "type": "string",
            "description": "Filtra clubes pela cidade (busca exata, case-insensitive).",
            "example": "Brasília"
          }
        }
      },
      "PaginationDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "Número da página a ser retornada.",
            "minimum": 1,
            "default": 1
          },
          "limit": {
            "type": "number",
            "description": "Número de itens por página.",
            "minimum": 1,
            "maximum": 100,
            "default": 10
          }
        }
      },
      "ClubDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do clube.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "name": {
            "type": "string",
            "description": "Nome oficial do clube.",
            "example": "Clube de Debate Oradores do Amanhã"
          },
          "city": {
            "type": "string",
            "description": "Cidade de localização do clube.",
            "example": "Brasília"
          },
          "state": {
            "type": "string",
            "description": "Estado de localização do clube.",
            "example": "DF"
          },
          "principalId": {
            "type": "string",
            "description": "ID do usuário que é o diretor do clube (principal).",
            "format": "uuid",
            "example": "f1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "createdAt": {
            "type": "string",
            "description": "Data de criação do clube.",
            "format": "date-time",
            "example": "2023-01-15T10:00:00Z"
          },
          "corum": {
            "type": "number",
            "description": "Quantidade de membros ativos no clube.",
            "example": 25
          }
        },
        "required": [
          "id",
          "name",
          "city",
          "state",
          "principalId",
          "createdAt",
          "corum"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "number",
            "example": 3
          },
          "total": {
            "type": "number",
            "example": 250
          },
          "limit": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "totalPages",
          "total",
          "limit",
          "page"
        ]
      },
      "PaginatedClubDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClubDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateClubInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do novo clube.",
            "example": "Clube de Debate Oradores do Amanhã",
            "minLength": 3
          },
          "city": {
            "type": "string",
            "description": "Cidade onde o clube está localizado.",
            "example": "Brasília",
            "minLength": 3
          },
          "state": {
            "type": "string",
            "description": "Estado onde o clube está localizado.",
            "example": "DF",
            "maxLength": 2,
            "minLength": 2
          }
        },
        "required": [
          "name",
          "city",
          "state"
        ]
      },
      "CreateClubOutputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do novo clube.",
            "example": "Clube de Debate Oradores do Amanhã",
            "minLength": 3
          },
          "city": {
            "type": "string",
            "description": "Cidade onde o clube está localizado.",
            "example": "Brasília",
            "minLength": 3
          },
          "state": {
            "type": "string",
            "description": "Estado onde o clube está localizado.",
            "example": "DF",
            "maxLength": 2,
            "minLength": 2
          },
          "accessToken": {
            "type": "string",
            "description": "Novo token de acesso JWT."
          },
          "refreshToken": {
            "type": "string",
            "description": "Novo token de atualização."
          }
        },
        "required": [
          "name",
          "city",
          "state",
          "accessToken",
          "refreshToken"
        ]
      },
...
      "EnrollmentRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da solicitação de matrícula.",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p6"
          },
          "status": {
            "type": "string",
            "description": "Status atual da solicitação.",
            "enum": [
              "REJECTED",
              "APPROVED",
              "PENDING"
            ],
            "example": "PENDING"
          },
          "clubId": {
            "type": "string",
            "description": "ID do clube ao qual a matrícula foi solicitada.",
            "format": "uuid",
            "example": "c1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p7"
          },
          "familyId": {
            "type": "string",
            "description": "ID da família que está solicitando a matrícula.",
            "format": "uuid",
            "example": "d1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p8"
          },
          "dependantId": {
            "type": "string",
            "description": "ID do dependente para o qual a matrícula se aplica.",
            "format": "uuid",
            "example": "e1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p9"
          },
          "requestedAt": {
            "type": "string",
            "description": "Data e hora em que a solicitação foi criada.",
            "format": "date-time",
            "example": "2025-07-20T14:00:00Z"
          },
          "resolvedAt": {
            "type": "string",
            "description": "Data e hora em que a solicitação foi resolvida (aprovada/rejeitada).",
            "format": "date-time",
            "nullable": true,
            "example": "2025-07-20T15:30:00Z"
          },
          "rejectionReason": {
            "type": "object",
            "description": "Motivo da rejeição, caso o status seja REJECTED.",
            "nullable": true,
            "example": "Vagas esgotadas para a faixa etária."
          }
        },
        "required": [
          "id",
          "status",
          "clubId",
          "familyId",
          "dependantId",
          "requestedAt",
          "resolvedAt",
          "rejectionReason"
        ]
      },
...
      "HolderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do titular.",
            "format": "uuid",
            "example": "b1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "firstName": {
            "type": "string",
            "description": "Nome do titular.",
            "example": "Maria Silva"
          },
          "lastName": {
            "type": "string",
            "description": "Nome do titular.",
            "example": "Maria Silva"
          },
          "email": {
            "type": "string",
            "description": "Email do titular.",
            "example": "maria.silva@email.com"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do titular.",
            "example": "(11) 98888-8888"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone"
        ]
      },
...
      "UpdateClubDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Novo nome do clube.",
            "example": "Clube de Oratória Avançada",
            "minLength": 3
          },
          "city": {
            "type": "string",
            "description": "Nova cidade de localização do clube.",
            "example": "São Paulo",
            "minLength": 3
          },
          "state": {
            "type": "string",
            "description": "Estado onde o clube está localizado.",
            "example": "DF",
            "maxLength": 2,
            "minLength": 2
          }
        },
        "required": [
          "state"
        ]
      },
      "ListPendingEnrollmentsOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da solicitação de matrícula.",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p6"
          },
          "status": {
            "type": "string",
            "description": "Status atual da solicitação.",
            "enum": [
              "REJECTED",
              "APPROVED",
              "PENDING"
            ],
            "example": "PENDING"
          },
          "clubId": {
            "type": "string",
            "description": "ID do clube ao qual a matrícula foi solicitada.",
            "format": "uuid",
            "example": "c1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p7"
          },
          "familyId": {
            "type": "string",
            "description": "ID da família que está solicitando a matrícula.",
            "format": "uuid",
            "example": "d1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p8"
          },
          "dependantId": {
            "type": "string",
            "description": "ID do dependente para o qual a matrícula se aplica.",
            "format": "uuid",
            "example": "e1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p9"
          },
          "dependantName": {
            "type": "string",
            "description": "Nome do dependente para o qual a matrícula se aplica.",
            "example": "Augustus Nicodemus"
          },
          "requestedAt": {
            "type": "string",
            "description": "Data e hora em que a solicitação foi criada.",
            "format": "date-time",
            "example": "2025-07-20T14:00:00Z"
          },
          "resolvedAt": {
            "type": "string",
            "description": "Data e hora em que a solicitação foi resolvida (aprovada/rejeitada).",
            "format": "date-time",
            "nullable": true,
            "example": "2025-07-20T15:30:00Z"
          },
          "rejectionReason": {
            "type": "object",
            "description": "Motivo da rejeição, caso o status seja REJECTED.",
            "nullable": true,
            "example": "Vagas esgotadas para a faixa etária."
          }
        },
        "required": [
          "id",
          "status",
          "clubId",
          "familyId",
          "dependantId",
          "dependantName",
          "requestedAt",
          "resolvedAt",
          "rejectionReason"
        ]
      },
      "RejectEnrollmentDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Motivo detalhado para a rejeição da matrícula. Mínimo de 10 caracteres.",
            "example": "Infelizmente, todas as vagas para esta faixa etária já foram preenchidas.",
            "minLength": 10
          }
        },
        "required": [
          "reason"
        ]
      },
      "ClubMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do membro.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "lastName": {
            "type": "string",
            "description": "Sobrenome do membro.",
            "example": "Silva"
          },
          "firstName": {
            "type": "string",
            "description": "Nome do membro.",
            "example": "João"
          },
          "email": {
            "type": "string",
            "description": "Email do membro.",
            "example": "joao.silva@email.com"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do membro.",
            "example": "(11) 99999-9999"
          },
          "holder": {
            "description": "Dados do titular.",
            "allOf": [
              {
                "$ref": "#/components/schemas/HolderDto"
              }
            ]
          },
          "memberSince": {
            "type": "string",
            "description": "Data de início da associação.",
            "format": "date-time",
            "example": "2023-01-15T10:00:00Z"
          },
          "birthDate": {
            "type": "string",
            "description": "Data de nascimento do membro.",
            "format": "date-time",
            "example": "1990-01-15T00:00:00Z"
          },
          "sex": {
            "type": "string",
            "description": "Sexo do membro.",
            "enum": [
              "FEMALE",
              "MALE"
            ],
            "example": "MALE"
          }
        },
        "required": [
          "id",
          "lastName",
          "firstName",
          "email",
          "phone",
          "holder",
          "memberSince",
          "birthDate",
          "sex"
        ]
      }
    }
  }
}